Renames daily_forecast_service to forecast_service so it's more apparent that it can be used for getting a forecast not just a daily forecast
